Adds metadata file and support for creating the Express subdirectory in S3 for PHP, and moves those files.
