+++ .doc_gen/metadata/s3_metadata.yaml
@@ -3481,32 +3481,6 @@ s3_Scenario_ConditionalRequests:
                 - python.example_code.s3.S3ConditionalRequests.wrapper
   services:
     s3: {GetObject, PutObject, CopyObject}
-s3_Scenario_ExpressBasics:
-  title: Learn the basics of Amazon S3 Express One Zone with an &AWS; SDK
-  title_abbrev: Learn the basics of S3 Express One Zone
-  synopsis_list:
-    - Set up a VPC and VPC Endpoint
-    - Set up the S3 Express Policies, Roles, and User to work with S3 Express buckets
-    - Create two S3 Clients
-    - Create two buckets
-    - Create an object and copy it over
-    - Demonstrate performance difference
-    - Populate the buckets to show the lexicographical difference
-    - Prompt the user to see if they want to clean up the resources
-  category: Basics
-  languages:
-    PHP:
-      versions:
-        - sdk_version: 3
-          github: php/example_code/s3
-          sdkguide:
-          excerpts:
-            - description:
-              snippet_tags:
-                - php.example_code.s3.ExpressBasics
-                - php.example_code.s3.service.S3Service
-  services:
-    s3: {CreateVpc, DescribeRouteTables, CreateVpcEndpoint, CreateBucket, CopyObject, GetObject, PutObject, ListObjects, DeleteObject, DeleteBucket, DeleteVpcEndpoint, DeleteVpc}
 s3_Scenario_DownloadS3Directory:
   title: Download S3 'directories' from an &S3long; (&S3;) bucket
   title_abbrev: Download S3 'directories'
