+++ php/example_code/s3/README.md
@@ -39,7 +39,6 @@ For prerequisites, see the [README](../../README.md#Prerequisites) in the `php` 
 Code examples that show you how to perform the essential operations within a service.
 
 - [Learn the basics](GettingStartedWithS3.php)
-- [Learn the basics of S3 Express One Zone](S3ExpressBasics.php)
 
 
 ### Single actions
@@ -96,26 +95,6 @@ This example shows you how to do the following:
 
 <!--custom.basics.s3_Scenario_GettingStarted.start-->
 <!--custom.basics.s3_Scenario_GettingStarted.end-->
-
-#### Learn the basics of S3 Express One Zone
-
-This example shows you how to do the following:
-
-- Set up a VPC and VPC Endpoint
-- Set up the S3 Express Policies, Roles, and User to work with S3 Express buckets
-- Create two S3 Clients
-- Create two buckets
-- Create an object and copy it over
-- Demonstrate performance difference
-- Populate the buckets to show the lexicographical difference
-- Prompt the user to see if they want to clean up the resources
-
-<!--custom.basic_prereqs.s3_Scenario_ExpressBasics.start-->
-<!--custom.basic_prereqs.s3_Scenario_ExpressBasics.end-->
-
-
-<!--custom.basics.s3_Scenario_ExpressBasics.start-->
-<!--custom.basics.s3_Scenario_ExpressBasics.end-->
 
 
 #### Create a presigned URL
